+++ src/integration-test/java/com/commercetools/sync/integration/commons/utils/CategoryITUtils.java
@@ -35,6 +35,8 @@ import java.util.stream.Collectors;
 
 import static com.commercetools.sync.integration.commons.utils.ITUtils.createCustomFieldsJsonMap;
 import static com.commercetools.sync.integration.commons.utils.ITUtils.createTypeIfNotAlreadyExisting;
+import static com.commercetools.sync.commons.utils.ResourceIdentifierUtils.toResourceIdentifierIfNotNull;
+
 import static io.sphere.sdk.utils.CompletableFutureUtils.listOfFuturesToFutureOfList;
 import static java.lang.String.format;
 
@@ -61,7 +63,8 @@ public final class CategoryITUtils {
             final String key = format("key%s", i + 1);
             final String orderHint = format("0.%s", i + 1);
             final CategoryDraft categoryDraft = CategoryDraftBuilder.of(name, slug)
-                                                                    .parent(parentCategory)
+                                                                    .parent(
+                                                                        toResourceIdentifierIfNotNull(parentCategory))
                                                                     .description(description)
                                                                     .key(key)
                                                                     .orderHint(orderHint)
@@ -130,7 +133,7 @@ public final class CategoryITUtils {
                 .of(LocalizedString.of(Locale.ENGLISH, categoryName),
                     LocalizedString.of(Locale.ENGLISH, categoryName))
                 .key(categoryName)
-                .parent(parent)
+                .parent(toResourceIdentifierIfNotNull(parent))
                 .custom(CustomFieldsDraft.ofTypeKeyAndJson(OLD_CATEGORY_CUSTOM_TYPE_KEY, createCustomFieldsJsonMap()))
                 .orderHint("sameOrderHint")
                 .build();