+++ src/main/java/com/commercetools/sync/commons/exceptions/BuildUpdateActionException.java
@@ -8,6 +8,10 @@ public class BuildUpdateActionException extends Exception {
         super(message);
     }
 
+    public BuildUpdateActionException(@Nonnull final Throwable cause) {
+        super(cause);
+    }
+
     public BuildUpdateActionException(@Nonnull final String message, @Nonnull final Throwable cause) {
         super(message, cause);
     }