+++ src/main/java/com/commercetools/sync/commons/helpers/BaseReferenceResolver.java
@@ -24,9 +24,10 @@ public abstract class BaseReferenceResolver<T, S extends BaseSyncOptions> {
     private static final String UUID_NOT_ALLOWED = "Found a UUID in the id field. Expecting a key without a UUID value."
         + " If you want to allow UUID values for reference keys, please use the allowUuidKeys(true) option in the"
         + " sync options.";
-    private static final String UNSET_ID_FIELD = "Reference 'id' field value is blank (null/empty).";
     private static final String KEY_NOT_SET_ON_EXPANSION_OR_ID_FIELD = "Key is blank (null/empty) on both expanded"
         + " reference object and reference id field.";
+    public static final String BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER = "The value of 'id' field of the Resource"
+        + " Identifier is blank (null/empty).";
     private static final String UUID_REGEX = "[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}";
 
     protected S options;
@@ -68,7 +69,7 @@ public abstract class BaseReferenceResolver<T, S extends BaseSyncOptions> {
                                                          final boolean allowUuidKeys)
         throws ReferenceResolutionException {
         final String key = resourceIdentifier.getId();
-        validateKey(key, allowUuidKeys, UNSET_ID_FIELD);
+        validateKey(key, allowUuidKeys, BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER);
         return key;
     }