+++ src/test/java/com/commercetools/sync/inventories/InventorySyncTest.java
@@ -27,6 +27,7 @@ import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 
+import static com.commercetools.sync.commons.helpers.BaseReferenceResolver.BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER;
 import static com.commercetools.sync.inventories.InventorySyncMockUtils.getCompletionStageWithException;
 import static com.commercetools.sync.inventories.InventorySyncMockUtils.getMockChannelService;
 import static com.commercetools.sync.inventories.InventorySyncMockUtils.getMockInventoryEntry;
@@ -366,8 +367,8 @@ public class InventorySyncTest {
         assertThat(errorCallBackMessages).isNotEmpty();
         assertThat(errorCallBackMessages.get(0)).contains(format("Failed to resolve references on"
             + " InventoryEntryDraft with SKU:'%s'. Reason: %s: Failed to resolve custom type reference on "
-            + "InventoryEntryDraft with SKU:'1000'. Reason: Reference 'id' field value is"
-            + " blank (null/empty).", SKU_1, ReferenceResolutionException.class.getCanonicalName()));
+            + "InventoryEntryDraft with SKU:'1000'. Reason: %s", SKU_1,
+            ReferenceResolutionException.class.getCanonicalName(), BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER));
         assertThat(errorCallBackExceptions).isNotEmpty();
         assertThat(errorCallBackExceptions.get(0)).isExactlyInstanceOf(CompletionException.class);
         assertThat(errorCallBackExceptions.get(0).getCause()).isExactlyInstanceOf(ReferenceResolutionException.class);